Poprawa ułożenia wszystkich modeli
