+++ src/Model/Admin/AdminUserDeleteModel.php
@@ -1,21 +1,17 @@
 <?php
 
-namespace App\Model\User;
+namespace App\Model\Admin;
 
-use App\Enums\UserRoles;
-
-class UserModel
+class AdminUserDeleteModel
 {
     private string $id;
     private bool $active;
     private bool $banned;
     private string $email;
     private string $firstname;
-    private string $lastname;
-    private int $dateCreated;
-
-    private array $roles = [];
     private bool $deleted;
+    private bool $declined;
+    private ?int $dateDeleted = null;
 
     /**
      * @param string $id
@@ -23,19 +19,18 @@ class UserModel
      * @param bool $banned
      * @param string $email
      * @param string $firstname
-     * @param string $lastname
-     * @param \DateTime $dateCreated
+     * @param bool $deleted
+     * @param bool $declined
      */
-    public function __construct(string $id, bool $active, bool $banned, string $email, string $firstname, string $lastname, \DateTime $dateCreated, bool $deleted)
+    public function __construct(string $id, bool $active, bool $banned, string $email, string $firstname, bool $deleted, bool $declined)
     {
         $this->id = $id;
         $this->active = $active;
         $this->banned = $banned;
         $this->email = $email;
         $this->firstname = $firstname;
-        $this->lastname = $lastname;
-        $this->dateCreated = $dateCreated->getTimestamp() * 1000;
         $this->deleted = $deleted;
+        $this->declined = $declined;
     }
 
     /**
@@ -119,58 +114,6 @@ class UserModel
     }
 
     /**
-     * @return string
-     */
-    public function getLastname(): string
-    {
-        return $this->lastname;
-    }
-
-    /**
-     * @param string $lastname
-     */
-    public function setLastname(string $lastname): void
-    {
-        $this->lastname = $lastname;
-    }
-
-    /**
-     * @return int
-     */
-    public function getDateCreated(): int
-    {
-        return $this->dateCreated;
-    }
-
-    /**
-     * @param \DateTime $dateCreated
-     */
-    public function setDateCreated(\DateTime $dateCreated): void
-    {
-        $this->dateCreated = $dateCreated->getTimestamp() * 1000;
-    }
-    /**
-     * @return string[]
-     */
-    public function getRoles(): array
-    {
-        return $this->roles;
-    }
-
-    /**
-     * @param array $roles
-     */
-    public function setRoles(array $roles): void
-    {
-        $this->roles = $roles;
-    }
-
-    public function addRole(UserRoles $role): void
-    {
-        $this->roles[] = $role->value;
-    }
-
-    /**
      * @return bool
      */
     public function isDeleted(): bool
@@ -184,6 +127,38 @@ class UserModel
     public function setDeleted(bool $deleted): void
     {
         $this->deleted = $deleted;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isDeclined(): bool
+    {
+        return $this->declined;
+    }
+
+    /**
+     * @param bool $declined
+     */
+    public function setDeclined(bool $declined): void
+    {
+        $this->declined = $declined;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getDateDeleted(): ?int
+    {
+        return $this->dateDeleted;
+    }
+
+    /**
+     * @param \DateTime $dateDeleted
+     */
+    public function setDateDeleted(\DateTime $dateDeleted): void
+    {
+        $this->dateDeleted = $dateDeleted->getTimestamp() * 1000;
     }
 
 }
