+++ src/Model/Admin/AdminUserModel.php
@@ -1,17 +1,21 @@
 <?php
 
-namespace App\Model\User;
+namespace App\Model\Admin;
 
-class UserDeleteModel
+use App\Enums\UserRoles;
+
+class AdminUserModel
 {
     private string $id;
     private bool $active;
     private bool $banned;
     private string $email;
     private string $firstname;
+    private string $lastname;
+    private int $dateCreated;
+
+    private array $roles = [];
     private bool $deleted;
-    private bool $declined;
-    private ?int $dateDeleted = null;
 
     /**
      * @param string $id
@@ -19,18 +23,19 @@ class UserDeleteModel
      * @param bool $banned
      * @param string $email
      * @param string $firstname
-     * @param bool $deleted
-     * @param bool $declined
+     * @param string $lastname
+     * @param \DateTime $dateCreated
      */
-    public function __construct(string $id, bool $active, bool $banned, string $email, string $firstname, bool $deleted, bool $declined)
+    public function __construct(string $id, bool $active, bool $banned, string $email, string $firstname, string $lastname, \DateTime $dateCreated, bool $deleted)
     {
         $this->id = $id;
         $this->active = $active;
         $this->banned = $banned;
         $this->email = $email;
         $this->firstname = $firstname;
+        $this->lastname = $lastname;
+        $this->dateCreated = $dateCreated->getTimestamp() * 1000;
         $this->deleted = $deleted;
-        $this->declined = $declined;
     }
 
     /**
@@ -114,6 +119,58 @@ class UserDeleteModel
     }
 
     /**
+     * @return string
+     */
+    public function getLastname(): string
+    {
+        return $this->lastname;
+    }
+
+    /**
+     * @param string $lastname
+     */
+    public function setLastname(string $lastname): void
+    {
+        $this->lastname = $lastname;
+    }
+
+    /**
+     * @return int
+     */
+    public function getDateCreated(): int
+    {
+        return $this->dateCreated;
+    }
+
+    /**
+     * @param \DateTime $dateCreated
+     */
+    public function setDateCreated(\DateTime $dateCreated): void
+    {
+        $this->dateCreated = $dateCreated->getTimestamp() * 1000;
+    }
+    /**
+     * @return string[]
+     */
+    public function getRoles(): array
+    {
+        return $this->roles;
+    }
+
+    /**
+     * @param array $roles
+     */
+    public function setRoles(array $roles): void
+    {
+        $this->roles = $roles;
+    }
+
+    public function addRole(UserRoles $role): void
+    {
+        $this->roles[] = $role->value;
+    }
+
+    /**
      * @return bool
      */
     public function isDeleted(): bool
@@ -127,38 +184,6 @@ class UserDeleteModel
     public function setDeleted(bool $deleted): void
     {
         $this->deleted = $deleted;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isDeclined(): bool
-    {
-        return $this->declined;
-    }
-
-    /**
-     * @param bool $declined
-     */
-    public function setDeclined(bool $declined): void
-    {
-        $this->declined = $declined;
-    }
-
-    /**
-     * @return int|null
-     */
-    public function getDateDeleted(): ?int
-    {
-        return $this->dateDeleted;
-    }
-
-    /**
-     * @param \DateTime $dateDeleted
-     */
-    public function setDateDeleted(\DateTime $dateDeleted): void
-    {
-        $this->dateDeleted = $dateDeleted->getTimestamp() * 1000;
     }
 
 }
