+++ src/ValueGenerator/BuildAudiobookCommentTreeGenerator.php
@@ -4,7 +4,7 @@ namespace App\ValueGenerator;
 
 use App\Entity\User;
 use App\Model\Common\AudiobookCommentsModel;
-use App\Model\User\UserAudiobookCommentModel;
+use App\Model\Common\AudiobookCommentModel;
 use App\Repository\AudiobookUserCommentLikeRepository;
 use App\Repository\AudiobookUserCommentRepository;
 use Symfony\Component\Uid\Uuid;
@@ -72,7 +72,7 @@ class BuildAudiobookCommentTreeGenerator implements ValueGeneratorInterface
                     "deleted" => false
                 ]);
 
-                $userModel = new UserAudiobookCommentModel($audiobookParentUser->getUserInformation()->getEmail(), $audiobookParentUser->getUserInformation()->getFirstname());
+                $userModel = new AudiobookCommentModel($audiobookParentUser->getUserInformation()->getEmail(), $audiobookParentUser->getUserInformation()->getFirstname());
 
                 $child = new AudiobookCommentsModel(
                     $userModel,
